🎨 Ajout tests
